test(NODE-4338): refactor auth prose tests
